+++ custom_components/energa_mobile/api.py
@@ -1,4 +1,4 @@
-"""API interface for Energa Mobile v3.6.0-beta.15."""
+"""API interface for Energa Mobile v3.6.0-beta.16."""
 import logging
 import aiohttp
 from datetime import datetime